@@ -14,33 +14,34 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.0+2
-  mask_text_input_formatter: ^1.0.2
-  flutter_icons: ^0.3.1
-  intl: ^0.16.0
-  shared_preferences: ^0.5.4+3
-  progress_dialog: ^1.2.0
-  google_maps_flutter: ^0.5.21+11
-  url_launcher: ^5.2.5
-  image_picker: 0.6.1+11
-  cached_network_image: ^1.1.3
-  expandable: ^3.0.1
+  font_awesome_flutter: 9.1.0
+  http: 0.13.4
+  mask_text_input_formatter: 2.0.0
+  flutter_icons: 1.1.0 
+  intl: 0.17.0
+  shared_preferences: 2.0.8
+  future_progress_dialog: 0.2.0
+  google_maps_flutter: 2.0.11
+  url_launcher: 6.0.12
+  image_picker: 0.8.4+2
+  cached_network_image: 3.1.0 
+  expandable: 5.0.1 
   
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: 1.0.3 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: "^0.7.3"
-  flutter_native_splash: ^0.1.9
+  flutter_launcher_icons: 0.9.2
+  flutter_native_splash: 1.2.4
 
 
 flutter_icons:
