@@ -1,45 +1,44 @@
 class CampaignModel {
-  final int id;
+  final int? id;
   final String name;
   final DateTime start;
-  final DateTime end;
+  final DateTime? end;
   final String description;
-  final String itemTypeTagName;
-  final String purposeTagName;
-  final String itemTypeTagColor;
-  final String purposeTagColor;
-  final double lat;
-  final double lng;
-  final int zipcode;
-  final String street;
-  final String city;
-  final int number;
-  final String state;
-  final String adonatorName;
-  final String adonatorEmail;
-  final String photoUrl;
-  final int campaignId;
+  final String? itemTypeTagName;
+  final String? purposeTagName;
+  final String? itemTypeTagColor;
+  final String? purposeTagColor;
+  final double? lat;
+  final double? lng;
+  final int? zipcode;
+  final String? street;
+  final String? city;
+  final int? number;
+  final String? state;
+  final String? adonatorName;
+  final String? adonatorEmail;
+  final String? photoUrl;
+  final int? campaignId;
 
-  const CampaignModel({
-    this.id,
-    this.name,
-    this.start,
-    this.end,
-    this.description,
-    this.itemTypeTagName,
-    this.purposeTagName,
-    this.itemTypeTagColor,
-    this.purposeTagColor,
-    this.lat,
-    this.lng,
-    this.zipcode,
-    this.street,
-    this.city,
-    this.number,
-    this.state,
-    this.adonatorName,
-    this.adonatorEmail,
-    this.photoUrl,
-    this.campaignId
-  });
+  const CampaignModel(
+      {this.id,
+      required this.name,
+      required this.start,
+      this.end,
+      required this.description,
+      this.itemTypeTagName,
+      this.purposeTagName,
+      this.itemTypeTagColor,
+      this.purposeTagColor,
+      this.lat,
+      this.lng,
+      this.zipcode,
+      this.street,
+      this.city,
+      this.number,
+      this.state,
+      this.adonatorName,
+      this.adonatorEmail,
+      this.photoUrl,
+      this.campaignId});
 }
