@@ -2,13 +2,9 @@ import 'package:flutter/material.dart';
 
 class ChipDesign extends StatelessWidget {
   final String label;
-  final Color color;
+  final Color? color;
 
-  const ChipDesign({
-    Key key,
-    this.label,
-    this.color
-  }) : super(key: key);
+  const ChipDesign({Key? key, this.label = '', this.color}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,10 +14,7 @@ class ChipDesign extends StatelessWidget {
         label: Text(
           this.label,
           style: TextStyle(
-            color: Colors.white,
-            fontSize: 10,
-            fontWeight: FontWeight.bold
-          ),
+              color: Colors.white, fontSize: 10, fontWeight: FontWeight.bold),
         ),
         backgroundColor: this.color,
         elevation: 4,
