@@ -18,8 +18,8 @@ class TagDropdowns extends StatefulWidget {
 }
 
 class TagDropdownsState extends State<TagDropdowns> {
-  String _purposeTag;
-  String _typeItemTag;
+  late String _purposeTag;
+  late String _typeItemTag;
   var purposeTagList;
   var itemTypeTagList;
 
@@ -29,27 +29,27 @@ class TagDropdownsState extends State<TagDropdowns> {
     _typeItemTag = widget.defaultItemTypeTagValue;
 
     purposeTagList = widget.tags
-      .where((value) => value.entries.toList()[3].value == "PURPOSE")
-      .map<DropdownMenuItem<String>>((value) {
-          var tag = value.entries.toList();
-          var dropdownMenuItem = DropdownMenuItem<String>(
-            value: tag[1].value,
-            child: Text(tag[1].value),
-          );
-          return dropdownMenuItem;
-        }
-      ).toList();
+        .where((value) => value.entries.toList()[3].value == "PURPOSE")
+        .map<DropdownMenuItem<String>>((value) {
+      var tag = value.entries.toList();
+      var dropdownMenuItem = DropdownMenuItem<String>(
+        value: tag[1].value,
+        child: Text(tag[1].value),
+      );
+      return dropdownMenuItem;
+    }).toList();
 
     itemTypeTagList = widget.tags
-      .where((value) => value.entries.toList()[3].value == "ITEM")
-      .map<DropdownMenuItem<String>>((value) {
+        .where((value) => value.entries.toList()[3].value == "ITEM")
+        .map<DropdownMenuItem<String>>(
+      (value) {
         var tag = value.entries.toList();
         var dropdownMenuItem = DropdownMenuItem<String>(
           value: tag[1].value,
           child: Text(tag[1].value),
         );
         return dropdownMenuItem;
-      }
+      },
     ).toList();
 
     super.initState();
@@ -57,49 +57,46 @@ class TagDropdownsState extends State<TagDropdowns> {
 
   @override
   Widget build(BuildContext context) {
-    return  Row(
-      children: <Widget>[
-        Expanded(
-          flex: 4,
-          child: Center(
-            child: DropdownButton(
-              hint: Text('O que doar', style: TextStyle(fontSize: 12, color: Colors.black)),
-              style: TextStyle(fontSize: 12, color: Colors.black),  
+    return Row(children: <Widget>[
+      Expanded(
+        flex: 4,
+        child: Center(
+          child: DropdownButton(
+              hint: Text('O que doar',
+                  style: TextStyle(fontSize: 12, color: Colors.black)),
+              style: TextStyle(fontSize: 12, color: Colors.black),
               items: itemTypeTagList,
               value: _typeItemTag,
-              onChanged: (value) {
-                _typeItemTag = value;
+              onChanged: (String? value) {
+                _typeItemTag = value!;
                 setState(() {
                   _typeItemTag = value;
                   widget.reference.defaultItemTypeTagValue = value;
                 });
-              }
-            ),
-          ),
+              }),
         ),
-        Expanded(
-          flex: 4,
-          child: Center(
-            child: DropdownButton(
-              hint: Text('Destino', style: TextStyle(fontSize: 12, color: Colors.black)),
-              style: TextStyle(fontSize: 12, color: Colors.black),
-              items: purposeTagList,
-              value: _purposeTag,
-              onChanged: (value) {
-                _purposeTag = value;
-                setState(() {
+      ),
+      Expanded(
+        flex: 4,
+        child: Center(
+          child: DropdownButton(
+            hint: Text('Destino',
+                style: TextStyle(fontSize: 12, color: Colors.black)),
+            style: TextStyle(fontSize: 12, color: Colors.black),
+            items: purposeTagList,
+            value: _purposeTag,
+            onChanged: (String? value) {
+              _purposeTag = value!;
+              setState(
+                () {
                   _purposeTag = value;
                   widget.reference.defaultPurpouseTagValue = value;
-                });
-              }
-            ),
+                },
+              );
+            },
           ),
         ),
-      ]
-    );
+      ),
+    ]);
   }
 }
-
-
-
-               
