@@ -1,35 +1,8 @@
 package com.example.adonate
 
-import android.os.Bundle
+import io.flutter.embedding.android.FlutterActivity;
 
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
-
-import android.os.Build
 import android.view.ViewTreeObserver
 import android.view.WindowManager
-class MainActivity: FlutterActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    val flutter_native_splash = true
-    var originalStatusBarColor = 0
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-        originalStatusBarColor = window.statusBarColor
-        window.statusBarColor = 0xff2d98e7.toInt()
-    }
-    val originalStatusBarColorFinal = originalStatusBarColor
 
-    GeneratedPluginRegistrant.registerWith(this)
-    val vto = flutterView.viewTreeObserver
-    vto.addOnGlobalLayoutListener(object : ViewTreeObserver.OnGlobalLayoutListener {
-      override fun onGlobalLayout() {
-        flutterView.viewTreeObserver.removeOnGlobalLayoutListener(this)
-        window.clearFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-          window.statusBarColor = originalStatusBarColorFinal
-        }
-      }
-    })
-
-  }
-}
+class MainActivity: FlutterActivity() {}
