@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:expandable/expandable.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:intl/intl.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -12,7 +11,7 @@ import 'package:adonate/shared/constants.dart';
 import 'package:adonate/model/CampaignModel.dart';
 
 class CampaignDetailActivity extends StatefulWidget {
-  CampaignDetailActivity({this.campaign});
+  CampaignDetailActivity({required this.campaign});
 
   final CampaignModel campaign;
 
@@ -33,8 +32,11 @@ class CampaignDetailActivityState extends State<CampaignDetailActivity> {
           child: Card(
             color: Colors.grey,
             child: CachedNetworkImage(
-              imageUrl: widget.campaign.photoUrl,
-              placeholder: (context, url) => Center(child: Container(color: Colors.grey,)),
+              imageUrl: widget.campaign.photoUrl ?? '',
+              placeholder: (context, url) => Center(
+                  child: Container(
+                color: Colors.grey,
+              )),
               errorWidget: (context, url, error) => Container(height: 0),
             ),
           ),
@@ -53,11 +55,11 @@ class CampaignDetailActivityState extends State<CampaignDetailActivity> {
 
     if (widget.campaign.end != null) {
       period =
-          'De ${formatter.format(widget.campaign.start)} \nAté ${formatter.format(widget.campaign.end)}';
+          'De ${formatter.format(widget.campaign.start)} \nAté ${formatter.format(widget.campaign.end!)}';
     }
 
     final CameraPosition _kGooglePlex = CameraPosition(
-      target: LatLng(widget.campaign.lat, widget.campaign.lng),
+      target: LatLng(widget.campaign.lat!, widget.campaign.lng!),
       zoom: 15.4746,
     );
 
@@ -65,7 +67,7 @@ class CampaignDetailActivityState extends State<CampaignDetailActivity> {
 
     final Marker marker = Marker(
       markerId: markerId,
-      position: LatLng(widget.campaign.lat, widget.campaign.lng),
+      position: LatLng(widget.campaign.lat!, widget.campaign.lng!),
       infoWindow: InfoWindow(
           title: widget.campaign.name, snippet: widget.campaign.adonatorName),
     );
@@ -74,133 +76,129 @@ class CampaignDetailActivityState extends State<CampaignDetailActivity> {
     markers.add(marker);
 
     return Scaffold(
-      backgroundColor: primaryColor,
-      appBar: AppBar(
-        iconTheme: IconThemeData(
-          color: Colors.white,
+        backgroundColor: primaryColor,
+        appBar: AppBar(
+          iconTheme: IconThemeData(
+            color: Colors.white,
+          ),
+          title: Text(
+            widget.campaign.name,
+            style: TextStyle(
+                fontSize: 20, fontWeight: FontWeight.bold, color: Colors.white),
+          ),
         ),
-        title: Text(
-          widget.campaign.name,
-          style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-              color: Colors.white),
-        ),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: <Widget> [
-            havePhoto(width),
-            Container(
-              width: width,
-              child: Card(
-                color: Colors.white,
-                elevation: 3,
-                child: Column(
-                  children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.only(right: 8.0, left: 8.0, top: 8),
-                      child: Text(
-                        period,
-                        style: TextStyle(fontSize: 16, color: Colors.grey),
+        body: SingleChildScrollView(
+            child: Column(children: <Widget>[
+          havePhoto(width),
+          Container(
+            width: width,
+            child: Card(
+              color: Colors.white,
+              elevation: 3,
+              child: Column(
+                children: <Widget>[
+                  Padding(
+                    padding:
+                        const EdgeInsets.only(right: 8.0, left: 8.0, top: 8),
+                    child: Text(
+                      period,
+                      style: TextStyle(fontSize: 16, color: Colors.grey),
+                    ),
+                  ),
+                  Padding(
+                    child: Text(
+                      'Criado por: ${widget.campaign.adonatorName}',
+                      style: TextStyle(fontSize: 16, color: Colors.grey),
+                    ),
+                    padding:
+                        const EdgeInsets.only(right: 8.0, left: 8.0, top: 8.0),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.all(8.0),
+                    child: ExpandablePanel(
+                      collapsed: Text(widget.campaign.description,
+                          style: TextStyle(color: Colors.grey),
+                          softWrap: true,
+                          maxLines: 2,
+                          overflow: TextOverflow.ellipsis),
+                      expanded: Text(
+                        widget.campaign.description,
+                        style: TextStyle(color: Colors.grey),
+                        softWrap: true,
+                      ),
+                      theme: ExpandableThemeData(
+                        tapHeaderToExpand: true,
+                        hasIcon: true,
                       ),
                     ),
-                    Padding(
-                      child: Text(
-                        'Criado por: ${widget.campaign.adonatorName}',
-                        style: TextStyle(fontSize: 16, color: Colors.grey),
-                      ),
-                      padding: const EdgeInsets.only(right: 8.0, left: 8.0, top: 8.0),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.all(8.0),
-                      child: ExpandablePanel(
-                          collapsed: Text(
-                            widget.campaign.description,
-                            style: TextStyle(color: Colors.grey),
-                            softWrap: true,
-                            maxLines: 2,
-                            overflow: TextOverflow.ellipsis
-                          ),
-                          expanded: Text(
-                            widget.campaign.description,
-                            style: TextStyle(color: Colors.grey),
-                            softWrap: true,
-                          ),
-                          tapHeaderToExpand: true,
-                          hasIcon: true,
-                        )
-                    ),
-                  ],
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                ),
+                  ),
+                ],
+                crossAxisAlignment: CrossAxisAlignment.start,
               ),
             ),
-            Container(
-              padding: EdgeInsets.all(16),
-              color: primaryColor,
-              child: Center(
+          ),
+          Container(
+            padding: EdgeInsets.all(16),
+            color: primaryColor,
+            child: Center(
                 child: Text(
-                  'Localização',
-                  style: TextStyle(fontSize: 20, color: Colors.white),
-                )
+              'Localização',
+              style: TextStyle(fontSize: 20, color: Colors.white),
+            )),
+          ),
+          Container(
+            height: 300,
+            width: width,
+            color: primaryColor,
+            child: Padding(
+              padding: const EdgeInsets.all(10.0),
+              child: GoogleMap(
+                mapType: MapType.terrain,
+                myLocationButtonEnabled: true,
+                myLocationEnabled: true,
+                mapToolbarEnabled: true,
+                markers: markers,
+                initialCameraPosition: _kGooglePlex,
+                compassEnabled: true,
+                onMapCreated: (GoogleMapController controller) {
+                  _controller.complete(controller);
+                },
               ),
             ),
-            Container(
-              height: 300,
-              width: width,
-              color: primaryColor,
-              child: Padding(
-                padding: const EdgeInsets.all(10.0),
-                child: GoogleMap(
-                  mapType: MapType.terrain,
-                  myLocationButtonEnabled: true,
-                  myLocationEnabled: true,
-                  mapToolbarEnabled: true,
-                  markers: markers,
-                  initialCameraPosition: _kGooglePlex,
-                  compassEnabled: true,
-                  onMapCreated: (GoogleMapController controller) {
-                    _controller.complete(controller);
-                  },
-                ),
-              ),
-            ),
-            Container(
-              width: width,
-              padding: const EdgeInsets.only(bottom: 18.0),
-              color: primaryColor,
-              child: ButtonTheme.bar(
-                child: ButtonBar(
-                  alignment: MainAxisAlignment.center,
-                  children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.only(right: 18.0),
-                      child: FloatingActionButton(
-                        heroTag: 'email',
-                        backgroundColor: Colors.orange,
-                        onPressed: _launchEmail,
-                        elevation: 3,
-                        child: Icon(Icons.email),
-                      ),
+          ),
+          Container(
+            width: width,
+            padding: const EdgeInsets.only(bottom: 18.0),
+            color: primaryColor,
+            child: ButtonBarTheme(
+              data: ButtonBarThemeData(alignment: MainAxisAlignment.center),
+              child: ButtonBar(
+                alignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  Padding(
+                    padding: const EdgeInsets.only(right: 18.0),
+                    child: FloatingActionButton(
+                      heroTag: 'email',
+                      backgroundColor: Colors.orange,
+                      onPressed: _launchEmail,
+                      elevation: 3,
+                      child: Icon(Icons.email),
                     ),
-                    Padding(
-                      padding: const EdgeInsets.only(left: 18.0),
-                      child: FloatingActionButton(
-                          heroTag: 'call',
-                          backgroundColor: Colors.orange,
-                          onPressed: _launchCellphone,
-                          elevation: 3,
-                          child: Icon(Icons.call)),
-                    )
-                  ],
-                ),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.only(left: 18.0),
+                    child: FloatingActionButton(
+                        heroTag: 'call',
+                        backgroundColor: Colors.orange,
+                        onPressed: _launchCellphone,
+                        elevation: 3,
+                        child: Icon(Icons.call)),
+                  )
+                ],
               ),
-            )
-          ]
-        )
-      )
-    );
+            ),
+          )
+        ])));
   }
 
   _launchEmail() async {
