@@ -9,7 +9,7 @@ class SharedPreferencesHelper {
 
   static get(key) async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    String value = prefs.getString(key);
+    String? value = prefs.getString(key);
     return value;
   }
 
